moved SingletonLoggerFactory to api.loggign and added tests
